legacy/requirements.txt
# AI Literature Analyzer Dependencies

# PDF处理
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# HTTP请求
requests>=2.31.0

# 配置文件处理
PyYAML>=6.0

# 数据处理
pathlib2>=2.3.7; python_version < "3.4"

# 可选依赖 (用于更好的文本处理)
# spacy>=3.7.0
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
